remove duplicate list contexts
