+++ packages/app/src/contexts/List/index.tsx
@@ -2,10 +2,13 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import type { AnyJson } from '@w3ux/types'
-import type { ListFormat } from 'library/PoolList/types'
 import { createContext, useContext, useState } from 'react'
 import { defaultContext } from './defaults'
-import type { ListContextInterface, ListProviderProps } from './types'
+import type {
+  ListContextInterface,
+  ListFormat,
+  ListProviderProps,
+} from './types'
 
 export const ListContext = createContext<ListContextInterface>(defaultContext)
