+++ packages/app/src/library/List/Selectable.tsx
@@ -1,11 +1,11 @@
 // Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+import { useList } from 'contexts/List'
 import { useUnstaking } from 'hooks/useUnstaking'
 import { useTranslation } from 'react-i18next'
 import { ButtonMonoInvert } from 'ui-buttons'
 import { SelectableWrapper } from '.'
-import { useList } from './context'
 import type { SelectableProps } from './types'
 
 export const Selectable = ({
