+++ packages/app/src/library/Nominations/types.ts
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import type { AnyJson } from '@w3ux/types'
+import type { ListFormat } from 'contexts/List/types'
 import type { Validator } from 'contexts/Validators/types'
-import type { ListFormat } from 'library/PoolList/types'
 
 export interface ManageNominationsInterface {
   addToSelected: (item: AnyJson) => void
