+++ packages/app/src/library/ValidatorList/Item.tsx
@@ -4,6 +4,7 @@
 import { faBars, faGlobe } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import type { AnyJson } from '@w3ux/types'
+import { useList } from 'contexts/List'
 import { useMenu } from 'contexts/Menu'
 import { usePlugins } from 'contexts/Plugins'
 import { EraPoints } from 'library/List/EraPoints'
@@ -23,7 +24,6 @@ import {
 } from 'ui-core/list'
 import { useOverlay } from 'ui-overlay'
 import { useValidators } from '../../contexts/Validators/ValidatorEntries'
-import { useList } from '../List/context'
 import { Blocked } from '../ListItem/Labels/Blocked'
 import { Commission } from '../ListItem/Labels/Commission'
 import { EraStatus } from '../ListItem/Labels/EraStatus'
