+++ packages/app/src/overlay/canvas/PoolMembers/Lists/Member.tsx
@@ -5,12 +5,12 @@ import { faBars, faShare, faUnlockAlt } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import type { AnyJson } from '@w3ux/types'
 import { useApi } from 'contexts/Api'
+import { useList } from 'contexts/List'
 import { useMenu } from 'contexts/Menu'
 import { useActivePool } from 'contexts/Pools/ActivePool'
 import { usePoolMembers } from 'contexts/Pools/PoolMembers'
 import { usePrompt } from 'contexts/Prompt'
 import { motion } from 'framer-motion'
-import { useList } from 'library/List/context'
 import { Identity } from 'library/ListItem/Labels/Identity'
 import { PoolMemberBonded } from 'library/ListItem/Labels/PoolMemberBonded'
 import { Select } from 'library/ListItem/Labels/Select'
