+++ packages/app/src/pages/Pools/types.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { ListFormat } from 'library/PoolList/types'
+import type { ListFormat } from 'contexts/List/types'
 import type { ActivePool } from 'types'
 
 export interface PoolAccountProps {
